Rename the SYSTEM module to SOCKET
This prevents a naming conflict with Lumen's SYSTEM module.
