+++ motor.l
@@ -59,7 +59,7 @@ ssize_t read(int fildes, void *buf, size_t nbyte);
 ssize_t write(int fildes, const void *buf, size_t nbyte);
 |)
 
-(require 'system)
+(require 'socket)
 
 (define cstr ffi.string)
 (define c ffi.C)
